+++ api-client-ruby/lib/daytona_api_client/models/list_secrets_response.rb
@@ -47,7 +47,7 @@ module DaytonaApiClient
     def self.openapi_types
       {
         :'items' => :'Array<Secret>',
-        :'total' => :'Float',
+        :'total' => :'Integer',
         :'next_cursor' => :'String'
       }
     end
