+++ api-client-ruby/lib/daytona_api_client/models/paginated_sandboxes_deprecated.rb
@@ -47,9 +47,9 @@ module DaytonaApiClient
     def self.openapi_types
       {
         :'items' => :'Array<Sandbox>',
-        :'total' => :'Float',
-        :'page' => :'Float',
-        :'total_pages' => :'Float'
+        :'total' => :'Integer',
+        :'page' => :'Integer',
+        :'total_pages' => :'Integer'
       }
     end
